Move useCases folder from /src/domain to /src
