+++ src/adapters/http/index.ts
@@ -5,13 +5,13 @@ import helmet from "helmet";
 import * as cookies from "cookie-parser";
 import { Config } from "../../config";
 import { Logger } from "../../domain/logger";
-import { ConfirmConsentUseCase } from "../../domain/useCases/ConfirmConsentUseCase";
-import { ClientListUseCase } from "../../domain/useCases/ClientListUseCase";
-import { AuthenticateUseCase } from "../../domain/useCases/AuthenticateUseCase";
-import { FindGrantUseCase } from "../../domain/useCases/FindGrantUseCases";
-import { ProcessInteractionUseCase } from "../../domain/useCases/ProcessInteractionUseCase";
-import { RemoveGrantUseCase } from "../../domain/useCases/RemoveGrantUseCase";
-import { AbortInteractionUseCase } from "../../domain/useCases/AbortInteractionUseCase";
+import { ConfirmConsentUseCase } from "../../useCases/ConfirmConsentUseCase";
+import { ClientListUseCase } from "../../useCases/ClientListUseCase";
+import { AuthenticateUseCase } from "../../useCases/AuthenticateUseCase";
+import { FindGrantUseCase } from "../../useCases/FindGrantUseCases";
+import { ProcessInteractionUseCase } from "../../useCases/ProcessInteractionUseCase";
+import { RemoveGrantUseCase } from "../../useCases/RemoveGrantUseCase";
+import { AbortInteractionUseCase } from "../../useCases/AbortInteractionUseCase";
 import { ClientService } from "../../domain/clients/ClientService";
 import { GrantService } from "../../domain/grants/GrantService";
 import { InteractionService } from "../../domain/interactions/InteractionService";
