+++ src/useCases/ConfirmConsentUseCase.ts
@@ -2,13 +2,13 @@ import * as crypto from "crypto";
 import { pipe } from "fp-ts/lib/function";
 import * as O from "fp-ts/Option";
 import * as TE from "fp-ts/TaskEither";
-import { GrantService } from "../grants/GrantService";
-import { Grant, GrantId } from "../grants/types";
-import { InteractionService } from "../interactions/InteractionService";
-import { Interaction, InteractionId } from "../interactions/types";
-import { Logger } from "../logger";
-import { DomainError, makeDomainError, Seconds } from "../types";
-import { fromTEOtoTE, show } from "../utils";
+import { GrantService } from "../domain/grants/GrantService";
+import { Grant, GrantId } from "../domain/grants/types";
+import { InteractionService } from "../domain/interactions/InteractionService";
+import { Interaction, InteractionId } from "../domain/interactions/types";
+import { Logger } from "../domain/logger";
+import { DomainError, makeDomainError, Seconds } from "../domain/types";
+import { fromTEOtoTE, show } from "../domain/utils";
 import { findValidGrant } from "./utils";
 
 type ConfirmConsentUseCaseError = DomainError;
