+++ src/useCases/RemoveGrantUseCase.ts
@@ -2,12 +2,12 @@ import * as TE from "fp-ts/TaskEither";
 import { constVoid, pipe } from "fp-ts/function";
 import { some } from "fp-ts/lib/Option";
 import * as RA from "fp-ts/ReadonlyArray";
-import { OrganizationId, ServiceId } from "../clients/types";
-import { GrantService } from "../grants/GrantService";
-import { Logger } from "../logger";
-import { DomainError, makeNotFoundError } from "../types";
-import { IdentityId } from "../identities/types";
-import { show } from "../utils";
+import { OrganizationId, ServiceId } from "../domain/clients/types";
+import { GrantService } from "../domain/grants/GrantService";
+import { Logger } from "../domain/logger";
+import { DomainError, makeNotFoundError } from "../domain/types";
+import { IdentityId } from "../domain/identities/types";
+import { show } from "../domain/utils";
 
 export type RemoveGrantUseCaseError = DomainError;
