+++ src/useCases/__tests__/FindGrantUseCases.test.ts
@@ -2,17 +2,17 @@ import * as mock from "jest-mock-extended";
 import * as O from "fp-ts/Option";
 import * as E from "fp-ts/Either";
 import * as TE from "fp-ts/TaskEither";
-import { GrantService } from "../../grants/GrantService";
-import { Logger } from "../../logger";
+import { GrantService } from "../../domain/grants/GrantService";
+import { Logger } from "../../domain/logger";
 import {
   FindGrantUseCase,
   InternalError,
   NotFound,
 } from "../FindGrantUseCases";
-import { client } from "../../clients/__tests__/data";
-import { identity } from "../../identities/__tests__/data";
-import { grant } from "../../grants/__tests__/data";
-import { makeDomainError } from "../../types";
+import { client } from "../../domain/clients/__tests__/data";
+import { identity } from "../../domain/identities/__tests__/data";
+import { grant } from "../../domain/grants/__tests__/data";
+import { makeDomainError } from "../../domain/types";
 
 const makeFindGrantUseCaseTest = () => {
   const logger = mock.mock<Logger>();
